requirements.txt
opencv-python>=4.5.0
matplotlib>=3.3.0
Pillow>=8.0.0
cryptography>=3.4.0
stegano>=0.10.0
numpy>=1.19.0
